// packages/panel-vendure/src/ui-extensions/orders-flow/generated-types.ts
export type Maybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Address = Node & {
  __typename?: 'Address';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  fullName: Maybe<Scalars['String']>;
  company: Maybe<Scalars['String']>;
  streetLine1: Scalars['String'];
  streetLine2: Maybe<Scalars['String']>;
  city: Maybe<Scalars['String']>;
  province: Maybe<Scalars['String']>;
  postalCode: Maybe<Scalars['String']>;
  country: Country;
  phoneNumber: Maybe<Scalars['String']>;
  defaultShippingAddress: Maybe<Scalars['Boolean']>;
  defaultBillingAddress: Maybe<Scalars['Boolean']>;
  customFields: Maybe<Scalars['JSON']>;
};
// ...
export type Asset = Node & {
  __typename?: 'Asset';
  tags: Array<Tag>;
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name: Scalars['String'];
  type: AssetType;
  fileSize: Scalars['Int'];
  mimeType: Scalars['String'];
  width: Scalars['Int'];
  height: Scalars['Int'];
  source: Scalars['String'];
  preview: Scalars['String'];
  focalPoint: Maybe<Coordinate>;
  customFields: Maybe<Scalars['JSON']>;
};
// ...
export type Channel = Node & {
  __typename?: 'Channel';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  code: Scalars['String'];
  token: Scalars['String'];
  defaultTaxZone: Maybe<Zone>;
  defaultShippingZone: Maybe<Zone>;
  defaultLanguageCode: LanguageCode;
  currencyCode: CurrencyCode;
  pricesIncludeTax: Scalars['Boolean'];
  customFields: Maybe<ChannelCustomFields>;
};

export type ChannelCustomFields = {
  __typename?: 'ChannelCustomFields';
  showOnMultitienda: Maybe<Scalars['Boolean']>;
  nombre: Maybe<Scalars['String']>;
  zoneStore: Maybe<Scalars['String']>;
  cityStore: Maybe<Scalars['String']>;
  provinceStore: Maybe<Scalars['String']>;
  description: Maybe<Scalars['String']>;
  geolocationStore: Maybe<Scalars['String']>;
  imgPortada: Maybe<Asset>;
  imgLogo: Maybe<Asset>;
  bgColorStore: Maybe<Scalars['String']>;
  storeEnabled: Maybe<Scalars['Boolean']>;
  messageStoreDisabled: Maybe<Scalars['String']>;
  emailStore: Maybe<Scalars['String']>;
  scheduleStore: Maybe<Scalars['String']>;
  phoneStore: Maybe<Scalars['String']>;
  telegramStore: Maybe<Scalars['String']>;
  whatsappStore: Maybe<Scalars['String']>;
  rrssStore: Maybe<Array<Scalars['String']>>;
  bannersAboutUs: Maybe<Array<Asset>>;
  imgSecondaryAboutUs: Maybe<Asset>;
  bodyAboutUs: Maybe<Scalars['String']>;
};
// ...
export type Customer = Node & {
  __typename?: 'Customer';
  groups: Array<CustomerGroup>;
  history: HistoryEntryList;
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  title: Maybe<Scalars['String']>;
  firstName: Scalars['String'];
  lastName: Scalars['String'];
  phoneNumber: Maybe<Scalars['String']>;
  emailAddress: Scalars['String'];
  addresses: Maybe<Array<Address>>;
  orders: OrderList;
  user: Maybe<User>;
  customFields: Maybe<Scalars['JSON']>;
};
// ...
export type CustomerGroup = Node & {
  __typename?: 'CustomerGroup';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name: Scalars['String'];
  customers: CustomerList;
  customFields: Maybe<Scalars['JSON']>;
};
// ...
export type Discount = {
  __typename?: 'Discount';
  adjustmentSource: Scalars['String'];
  type: AdjustmentType;
  description: Scalars['String'];
  amount: Scalars['Int'];
  amountWithTax: Scalars['Int'];
};
// ...
export type Fulfillment = Node & {
  __typename?: 'Fulfillment';
  nextStates: Array<Scalars['String']>;
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  orderItems: Array<OrderItem>;
  state: Scalars['String'];
  method: Scalars['String'];
  trackingCode: Maybe<Scalars['String']>;
  customFields: Maybe<Scalars['JSON']>;
};
// ...
export type HistoryEntryList = PaginatedList & {
  __typename?: 'HistoryEntryList';
  items: Array<HistoryEntry>;
  totalItems: Scalars['Int'];
};
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type OrderAddress = {
  __typename?: 'OrderAddress';
  fullName: Maybe<Scalars['String']>;
  company: Maybe<Scalars['String']>;
  streetLine1: Maybe<Scalars['String']>;
  streetLine2: Maybe<Scalars['String']>;
  city: Maybe<Scalars['String']>;
  province: Maybe<Scalars['String']>;
  postalCode: Maybe<Scalars['String']>;
  country: Maybe<Scalars['String']>;
  countryCode: Maybe<Scalars['String']>;
  phoneNumber: Maybe<Scalars['String']>;
  customFields: Maybe<Scalars['JSON']>;
};
// ...
export type OrderGroup = Node & {
  __typename?: 'OrderGroup';
  createdAt: Maybe<Scalars['DateTime']>;
  updatedAt: Maybe<Scalars['DateTime']>;
  orderPlacedAt: Maybe<Scalars['DateTime']>;
  id: Scalars['ID'];
  code: Maybe<Scalars['String']>;
  alias: Maybe<Scalars['String']>;
  state: Maybe<Scalars['String']>;
  active: Maybe<Scalars['Boolean']>;
  channel: Maybe<Channel>;
  customer: Maybe<Customer>;
  lines: Maybe<Array<Maybe<OrderLine>>>;
  linesGrouped: Maybe<Array<Maybe<OrderItemGrouped>>>;
  ordersConfirmed: Maybe<OrderList>;
  totalQuantity: Maybe<Scalars['Int']>;
  shipping: Maybe<Scalars['Int']>;
  shippingWithTax: Maybe<Scalars['Int']>;
  subTotal: Maybe<Scalars['Int']>;
  subTotalWithTax: Maybe<Scalars['Int']>;
  total: Maybe<Scalars['Int']>;
  totalWithTax: Maybe<Scalars['Int']>;
  ordersQuantity: Maybe<Scalars['Int']>;
  currencyCode: Maybe<Scalars['String']>;
  shippingLine: Maybe<ShippingLine>;
  shippingLines: Maybe<Array<Maybe<ShippingLine>>>;
  shippingAddress: Maybe<OrderAddress>;
  orders: Maybe<OrderList>;
  fulfillments: Maybe<Array<Fulfillment>>;
};
// ...
export type OrderItemGrouped = {
  __typename?: 'OrderItemGrouped';
  productorId: Maybe<Scalars['String']>;
  productorNombre: Maybe<Scalars['String']>;
  productoNombre: Scalars['String'];
  productoSku: Scalars['String'];
  shippingMethodCode: Maybe<Scalars['String']>;
  cantidad: Maybe<Scalars['Int']>;
  stockOnHand: Maybe<Scalars['Int']>;
  stockAllocated: Maybe<Scalars['Int']>;
  listPrice: Maybe<Scalars['Float']>;
  subTotallistPrice: Maybe<Scalars['Float']>;
};

/** Returned when the maximum order size limit has been reached. */
export type OrderLimitError = ErrorResult & {
  __typename?: 'OrderLimitError';
  errorCode: ErrorCode;
  message: Scalars['String'];
  maxItems: Scalars['Int'];
};

export type OrderLine = Node & {
  __typename?: 'OrderLine';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  productVariant: ProductVariant;
  featuredAsset: Maybe<Asset>;
  /** The price of a single unit, excluding tax and discounts */
  unitPrice: Scalars['Int'];
  /** The price of a single unit, including tax but excluding discounts */
  unitPriceWithTax: Scalars['Int'];
  /** Non-zero if the unitPrice has changed since it was initially added to Order */
  unitPriceChangeSinceAdded: Scalars['Int'];
  /** Non-zero if the unitPriceWithTax has changed since it was initially added to Order */
  unitPriceWithTaxChangeSinceAdded: Scalars['Int'];
  /**
   * The price of a single unit including discounts, excluding tax.
   *
   * If Order-level discounts have been applied, this will not be the
   * actual taxable unit price (see `proratedUnitPrice`), but is generally the
   * correct price to display to customers to avoid confusion
   * about the internal handling of distributed Order-level discounts.
   */
  discountedUnitPrice: Scalars['Int'];
  /** The price of a single unit including discounts and tax */
  discountedUnitPriceWithTax: Scalars['Int'];
  /**
   * The actual unit price, taking into account both item discounts _and_ prorated (proportionally-distributed)
   * Order-level discounts. This value is the true economic value of the OrderItem, and is used in tax
   * and refund calculations.
   */
  proratedUnitPrice: Scalars['Int'];
  /** The proratedUnitPrice including tax */
  proratedUnitPriceWithTax: Scalars['Int'];
  quantity: Scalars['Int'];
  items: Array<OrderItem>;
  taxRate: Scalars['Float'];
  /** The total price of the line excluding tax and discounts. */
  linePrice: Scalars['Int'];
  /** The total price of the line including tax but excluding discounts. */
  linePriceWithTax: Scalars['Int'];
  /** The price of the line including discounts, excluding tax */
  discountedLinePrice: Scalars['Int'];
  /** The price of the line including discounts and tax */
  discountedLinePriceWithTax: Scalars['Int'];
  /**
   * The actual line price, taking into account both item discounts _and_ prorated (proportionally-distributed)
   * Order-level discounts. This value is the true economic value of the OrderLine, and is used in tax
   * and refund calculations.
   */
  proratedLinePrice: Scalars['Int'];
  /** The proratedLinePrice including tax */
  proratedLinePriceWithTax: Scalars['Int'];
  /** The total tax on this line */
  lineTax: Scalars['Int'];
  discounts: Array<Discount>;
  taxLines: Array<TaxLine>;
  order: Order;
  customFields: Maybe<Scalars['JSON']>;
};

export type OrderLineInput = {
  orderLineId: Scalars['ID'];
  quantity: Scalars['Int'];
};

export type OrderList = PaginatedList & {
  __typename?: 'OrderList';
  items: Array<Order>;
  totalItems: Scalars['Int'];
};
// ...
export type PaginatedList = {
  items: Array<Node>;
  totalItems: Scalars['Int'];
};
// ...
export type ProductVariant = Node & {
  __typename?: 'ProductVariant';
  enabled: Scalars['Boolean'];
  trackInventory: GlobalFlag;
  stockOnHand: Scalars['Int'];
  stockAllocated: Scalars['Int'];
  outOfStockThreshold: Scalars['Int'];
  useGlobalOutOfStockThreshold: Scalars['Boolean'];
  stockMovements: StockMovementList;
  channels: Array<Channel>;
  id: Scalars['ID'];
  product: Product;
  productId: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  languageCode: LanguageCode;
  sku: Scalars['String'];
  name: Scalars['String'];
  featuredAsset: Maybe<Asset>;
  assets: Array<Asset>;
  price: Scalars['Int'];
  currencyCode: CurrencyCode;
  priceWithTax: Scalars['Int'];
  stockLevel: Scalars['String'];
  taxRateApplied: TaxRate;
  taxCategory: TaxCategory;
  options: Array<ProductOption>;
  facetValues: Array<FacetValue>;
  translations: Array<ProductVariantTranslation>;
  customFields: Maybe<Scalars['JSON']>;
};
// ...
export type ShippingLine = {
  __typename?: 'ShippingLine';
  shippingMethod: ShippingMethod;
  price: Scalars['Int'];
  priceWithTax: Scalars['Int'];
  discountedPrice: Scalars['Int'];
  discountedPriceWithTax: Scalars['Int'];
  discounts: Array<Discount>;
  id: Maybe<Scalars['ID']>;
};

export type ShippingMethod = Node & {
  __typename?: 'ShippingMethod';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  code: Scalars['String'];
  name: Scalars['String'];
  description: Scalars['String'];
  fulfillmentHandlerCode: Scalars['String'];
  checker: ConfigurableOperation;
  calculator: ConfigurableOperation;
  translations: Array<ShippingMethodTranslation>;
  customFields: Maybe<ShippingMethodCustomFields>;
};
// ...
export type TaxLine = {
  __typename?: 'TaxLine';
  description: Scalars['String'];
  taxRate: Scalars['Float'];
};
// ...
export type User = Node & {
  __typename?: 'User';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  identifier: Scalars['String'];
  verified: Scalars['Boolean'];
  roles: Array<Role>;
  lastLogin: Maybe<Scalars['DateTime']>;
  authenticationMethods: Array<AuthenticationMethod>;
  customFields: Maybe<Scalars['JSON']>;
};

export enum VirutalOrderState {
  NUEVOS = 'NUEVOS',
  EN_PREPARACION = 'EN_PREPARACION',
  EN_ENTREGA = 'EN_ENTREGA'
}

export type Zone = Node & {
  __typename?: 'Zone';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name: Scalars['String'];
  members: Array<Country>;
  customFields: Maybe<Scalars['JSON']>;
};